+++ js/templates/userPage.html
@@ -90,7 +90,7 @@
     <div class="bar navBar barFlush user-page-navigation-buttons positionBottom backgroundTransparent width950">
       <div class="navEnd pad15">
         <div class="ctrlGroup hide js-pageButtons js-userPageControls zIndex4">
-          <a class="btn btn-txt custCol-primary js-moderator">
+          <a class="btn btn-txt custCol-primary js-moderatorSettings">
             <span class="ion-ios-pricetag fontSize11 marginRight2 textOpacity1"></span>
             <% if(ob.page.profile.moderator){ %>
             <%= polyglot.t('EditModerator') %>
